+++ packages/ragbits-core/CHANGELOG.md
@@ -1,9 +1,11 @@
 # CHANGELOG
 
 ## Unreleased
-
+- Make the score in VectorStoreResult consistent (always bigger is better)
 - Add router option to LiteLLMEmbedder (#440)
 - Make LLM / Embedder APIs consistent (#463)
+- Fix: make unflatten_dict symmetric to flatten_dict (#461)
+- Cost and capabilities config for custom litellm models (#481)
 
 ## 0.12.0 (2025-03-25)
 - Allow Prompt class to accept the asynchronous response_parser. Change the signature of parse_response method.
@@ -11,7 +13,6 @@
 - Fix Qdrant vector store serialization (#419)
 
 ## 0.11.0 (2025-03-25)
-
 - Add HybridSearchVectorStore which can aggregate results from multiple VectorStores (#412)
 
 ## 0.10.2 (2025-03-21)